Database clean up
Initialize database at process start, and allow for an in-memory database type
to faciliate unit testing. As a result of initializing at launch, you no longer
need to pass a context to most db methods. Only functions that schedule a backup
need it because they must obtain a pointer to the JobScheduler service.

Also added a whole suite of unit tests around the DB functionality.