Use a padding to avoid overlap at the bottom of edge-to-edge screens instead of a placeholder view
In the case where there was no overlapping view, some devices would report a nav bar inset of 0
which when placed on the layout of the navigation bar placeholder view would cause it to
match the parent view because 0 is the value for 'match_parent'. Where relevant, the app now
sets padding on the relevant bottom view.